transport/ntcp2: add proper tcp listener support
Implement first version of the transport architecture, introducing
`TransportManager`, `Ntcp2Transport` and `Ntc2pListener` objects.

`TransportManager` is the high-level transport object responsible for
polling lower-level transport objects, namely NTCP2 and SSU2.

Each transport will implement `Transport` trait which provides a common
interfaces for the transports and allows `TransportManager` to opaquely
calls the respective transport object.

Transport listeners (only `Ntcp2Listener`) in this commit are
responsible for listening for incoming connections and accepting
and negotiating the inbound session. Once the session has been
negotiated, `TransportManager` will be notified, which in turn will
notifiy the higher-level protocols of the connection,
and the connection itself is spawned in the background