streaming: flesh out streaming protocol architecture
`StreamManager` is the high-level object which deals with inbound
and outbound streaming packets. Messages received from `SamSession`
are checked against the local records to verify whether the packet
belongs to a new stream or an existing stream. If the packet belongs
to an existing stream, `StreamManager` sends the packet to that
stream's handler which sends to the connected client

if the message belongs to a new stream (`SYN` flag set), `StreamManager`
forwards the stream to `StreamListener` which checks if there is an
active listener, either ephemeral (`STREAM ACCEPT`) or persistent
(`STREAM FORWARD`) and if so, creates a new `Stream` object and spawns
a task in the background

`Stream` represents an active i2p virtual stream and is directly
connected to both `StreamManager`, for messages received from the
network, and to the samv3 client, for sending messages to the network