tunnel: implement feedback mechanism for transit/tunnel pool managers
before a transit tunnel is finally accepted and its event loop is
started, it awaits feedback from `TunnelManager` whether the next hop
was dialed succesfully or not. If dial failed for the next hop, the
transit tunnel is destroyed as there is nothing to do and it makes no
sense to keep an inactive transit tunnel in the active transit tunnel
storage, decreasing the amount of active transit tunnels the router
could have

the same principle operates to tunnel pool and the tunnel build
messages it sends. If the next hop of a tunnel build message cannot be
dialed, the tunnel pool listener is notified of it immediately instead
of waiting the full 10 seconds for tunnel build timeout, allowing
tunnel pool to build replacement tunnels faster