session: Respond to ACK requests if no traffic is sent (#78)
This PR aims to close https://github.com/altonen/emissary/issues/39.

This PR introduces a `LOW_PRIORITY_RESPONSE_INTERVAL` of 1 second for
ACK requests. If no message traffic is sent within this time-frame, an
empty message will be scheduled to be sent to the router.

Future PRs will handle NextKey responses and "high priority" NS and NSR
responses (though this will be trivial, so I can include that in this PR
if you prefer).