cli/http: Use `.b32.i2p` hostnames in HTTP requests (#116)
when a `.i2p` hostname is used to connect to an eepsite, convert the
value of the `Host` field of the request to a `.b32.i2p` of that host,
if it exists in the address book.

introduce `AddressBook::resolve_b32()` which attempts to resolve a
`.i2p` hostname into its destination hash and it is found, the value is
cached in `AddressBook` in an LRU cache.

Resolves #108